third_party/blink/web_tests/platform/mac-mac12/external/wpt/css/css-values/clamp-length-serialize-expected.txt

This is a testharness.js-based test.
FAIL e.style['letter-spacing'] = "clamp(1px, 2px, 3px)" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, 3px)"
FAIL e.style['letter-spacing'] = "calc(clamp(1px, 2px, 3px))" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, 3px)"
FAIL e.style['letter-spacing'] = "clamp(1px, 2px, clamp(2px, 3px, 4px))" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, clamp(2px, 3px, 4px))"
FAIL e.style['letter-spacing'] = "calc(clamp(1px, 2px, clamp(2px, 3px, 4px)))" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, clamp(2px, 3px, 4px))"
Harness: the test ran to completion.
